Creating UI package. IAlert interface is used in the application, by the implementation of Alerts. Alerts in the application will be created by Alerts class
